0.2.0: Improved Grouping & Verbose Mode
Implemented a new verbose flag which prints information about the
currently-running tasks. Additionally, the grouping is now being
handled differently. When grouping is enabled, the `Child` handle
will be collected; and it's standard output/error piped and
channeled. This change will allow messages to be printed on-demand
as soon as they become available.