0.5.0: Input Disk Buffering; Error Handling; & More This release required a ton of work to accomplish, and about two weeks of my life, as I wanted to ensure that I wrote it as efficiently as possible. Additional performance optimizations will be performed in the future.
There are a significant number of changes in this release, but for
the end user this will mainly improve performance, drastically
reduce memory consumption, and fix the bug that causes the program
to hang when spawning a large number of processes. This bug was
caused by spawned children not automatically releasing their handles
as intended. Errors will now also be logged and re-printed at the
end of the program.

With this release, arguments will first be efficiently buffered
to the disk using a custom `DiskBufferWriter` structure that
works entirely with a fixed-size 8K byte array. The structures
is created from a `DiskBuffer` which transforms into either a
`DiskBufferReader` or a `DiskBufferWriter` depending upon whether
the `read()` or `write()` method was called. When the byte array is
full, the data will be written to the disk. This is significantly
more efficient than using the default BufWriter structure, consuming
far less syscalls.

Once all arguments have been written to the disk, the tokenizer
will generate tokens from the command argument if one was given,
but will now instantly replace instances of `JobTotal` and
`Number` tokens during the tokenizing phase, because this
information is known during initialization.

The end of the initialization phase is now marked by the creation
of an `InputIterator` structure, which contains an `InputBuffer`
with higher level information on an internal `DiskBufferReader`.
The iterator is wrapped inside of an atomic reference-counted
mutex and shared among threads. Threads will continue to consume
from the iterator until the iterator returns `None`.

Finally, the entirety of the codebase was refactored into smaller
modules, which took a large amount of time, but the result is
code being easier to read and manage.